+++ src/app/components/forms/input/radio/radio-group.component.spec.ts
@@ -30,9 +30,11 @@ describe(`RadioGroupComponent:`, () =>
     beforeEach(async(() =>
     {
         TestBed.configureTestingModule({
-            declarations: [RadioInputComponent,
-                           RadioGroupComponent,
-                           RadioGroupTestComponent]
+            declarations: [
+                RadioInputComponent,
+                RadioGroupComponent,
+                RadioGroupTestComponent
+            ]
         }).compileComponents();
     }));